requirements.txt
Django~=3.2.3
requests~=2.22.0
djangorestframework~=3.12.4
Pillow~=8.2.0
celery==5.0.5
django_celery_beat
django-cors-headers
django_celery_results
djangorestframework-simplejwt
redis
python-telegram-bot